requirements.txt
asgiref              ==3.8.1
attrs                ==24.2.0
autobahn             ==24.4.2
Automat              ==24.8.1
certifi              ==2024.8.30
cffi                 ==1.17.1
channels             ==3.0.5
charset-normalizer   ==3.3.2
constantly           ==23.10.4
cryptography         ==43.0.1
daphne               ==3.0.2
Django               ==5.1.1
django-allauth       ==64.2.1
django-widget-tweaks ==1.5.0
djangorestframework  ==3.15.2
hyperlink            ==21.0.0
idna                 ==3.8
incremental          ==24.7.2
jwt                  ==1.3.1
pillow               ==10.4.0
pip                  ==23.0.1
pyasn1               ==0.6.0
pyasn1_modules       ==0.4.0
pycparser            ==2.22
pyOpenSSL            ==24.2.1
requests             ==2.32.3
service-identity     ==24.1.0
setuptools           ==66.1.1
sqlparse             ==0.5.1
Twisted              ==24.7.0
txaio                ==23.1.1
typing_extensions    ==4.12.2
urllib3              ==2.2.2
zope.interface       ==7.0.3
